feat: allow posting to multiple platforms??
this warrants more discussion - because the requirements for multiple platforms are different.

Maybe this warrants another discussion about the nature of a "post" as well - do we want the same "single" post to be "posted on multiple platforms"? Is this how we think of social media posts? Leaving this as a commit in case we want to revert it in the future.